# packages/api-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Self Deepsearch Platform API
  version: 0.1.0
  description: Release A contract baseline. Automatic collection is not enabled.
servers:
  - url: http://127.0.0.1:8080
paths:
...
  /api/v1/site/sitemap:
    get:
      operationId: getSitemapEntries
      description: Returns only currently published URLs, optionally filtered for split sitemaps.
      parameters:
        - in: query
          name: entity_type
          schema: { enum: [all, work, performer, studio], default: all }
      responses:
        "200":
          description: Published sitemap entries
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SitemapResponse" }
  /api/v1/site/redirect:
    get:
      operationId: resolvePublicRedirect
      description: Resolves only active, internal, permanent catalog redirects.
      parameters:
        - in: query
          name: path
          required: true
          schema: { type: string, maxLength: 300, pattern: "^/(works|performers|studios)/" }
      responses:
        "200":
          description: Safe internal redirect target
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RedirectTargetResponse" }
        "400": { description: Invalid path }
        "404": { description: No active redirect }
  /api/v1/search/works:
    get:
      operationId: searchWorks
      description: Searches published work documents by code first, then title, performer names and aliases, and studio name.
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
            minLength: 2
            maxLength: 100
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Published works matching the query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkSearchResponse"
        "400":
          description: Invalid code query
  /api/v1/works/popular:
    get:
      operationId: listPopularWorks
      description: Lists published works ranked by hourly aggregate page views without storing anonymous identifiers.
      parameters:
        - in: query
          name: range
          schema: { enum: [7d, 30d, all], default: 30d }
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Popular published works
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ItemListResponse" }
        "400": { description: Invalid range }
  /api/v1/works/latest:
    get:
      operationId: listLatestReleasedWorks
      description: Lists published works by actual release date; records without a release date are excluded.
      parameters:
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Latest released published works
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ItemListResponse" }
        "400": { description: Invalid cursor }
  /api/v1/works/recent:
    get:
      operationId: listRecentlyAddedWorks
      description: Lists published works by their first public publication time rather than release date.
      parameters:
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Recently published catalog entries
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ItemListResponse" }
  /api/v1/works/{slug}:
    get:
      operationId: getPublishedWork
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
            pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
      responses:
        "200":
          description: Published work detail. This read is side-effect free.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkDetailResponse"
        "404":
          description: Work not found or not published
  /api/v1/performers:
    get:
      operationId: listPublishedPerformers
      parameters:
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Recently published performer profiles
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ItemListResponse" }
  /api/v1/performers/{slug}:
    get:
      operationId: getPublishedPerformer
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string, pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$" }
      responses:
        "200":
          description: Published, adult-verified performer profile and related published works
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PerformerDetailResponse" }
        "404": { description: Performer not found or not published }
  /api/v1/studios:
    get:
      operationId: listPublishedStudios
      parameters:
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Published studios
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StudioListResponse" }
  /api/v1/studios/{slug}:
    get:
      operationId: getPublishedStudio
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string, pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$" }
      responses:
        "200":
          description: Published studio and related published works
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StudioDetailResponse" }
        "404": { description: Studio not found or not published }
  /api/v1/auth/signup/code:
    post:
      operationId: requestSignupCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthCodeRequest"
      responses:
        "202":
          description: Uniform response whether or not the email can be registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Email format or human verification is invalid
  /api/v1/auth/signup/verify:
    post:
      operationId: verifySignup
      description: Password work shares a process-wide bounded queue. Overload returns AUTH_BUSY with Retry-After, without consuming the code or creating a session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: Account and initial server-side session created atomically
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "503":
          description: AUTH_BUSY when the password queue is full or times out; retry without credential changes
          headers:
            Retry-After:
              schema: { type: string, enum: ["1"] }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/auth/invitations/accept:
    post:
      operationId: acceptInvitation
      description: Prechecks the single-use invitation code and daily quota before bounded password work, then rechecks and atomically consumes the code, creates the invited account and starts a session. Precheck does not consume quota.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvitationAcceptRequest"
      responses:
        "201":
          description: Invited account and initial server-side session created atomically
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "400": { description: Invalid or expired invitation code or password }
        "409": { description: Invitation was already accepted, revoked, expired or the email already has an account }
        "429": { description: Invitation acceptance quota exceeded }
        "503":
          description: AUTH_BUSY when the password queue is full or times out; invitation remains unconsumed
          headers:
            Retry-After:
              schema: { type: string, enum: ["1"] }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/auth/login:
    post:
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login succeeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "401":
          description: Generic invalid credential response
        "503":
          description: AUTH_BUSY when the password queue is full or times out; no session issued and not a wrong-password result
          headers:
            Retry-After:
              schema: { type: string, enum: ["1"] }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/auth/logout:
    post:
      operationId: logout
      description: Idempotently revokes the current session. Only confirmed revocation or an absent session clears the session and recent-auth cookies. Dependency failures leave cookies unchanged so the caller can retry; they never imply successful logout.
      responses:
        "204":
          description: Current session revoked or already absent
        "503":
          description: Session revocation could not be confirmed; cookies remain unchanged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/auth/reauth:
    post:
      operationId: reauthenticate
      description: Verifies the current session password and issues a five-minute HttpOnly confirmation cookie for high-risk operator mutations. The token is bound to the current session and is never returned in JSON.
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReauthenticateRequest"
      responses:
        "200":
          description: Recent password confirmation accepted
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401": { description: Session absent or password invalid }
        "429": { description: Reauthentication attempts are rate limited }
        "503":
          description: AUTH_BUSY when the password queue is full or times out; no confirmation cookie issued
          headers:
            Retry-After:
              schema: { type: string, enum: ["1"] }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/auth/password/code:
    post:
      operationId: requestPasswordResetCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthCodeRequest"
      responses:
        "202":
          description: Uniform response whether or not an active account exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /api/v1/account/close/code:
    post:
      operationId: requestAccountCloseCode
      security:
        - sessionCookie: []
      responses:
        "202":
          description: Account-close code queued for the authenticated email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  securitySchemes:
    metricsBearer:
      type: http
      scheme: bearer
    sessionCookie:
      type: apiKey
      in: cookie
      name: sd_session
    recentAuthCookie:
      type: apiKey
      in: cookie
      name: sd_reauth
  parameters:
    Cursor:
      in: query
      name: cursor
      required: false
      description: Opaque keyset cursor returned by the same route and query.
      schema: { type: string, maxLength: 256 }
...
  schemas:
...
    SitemapResponse:
      type: object
      required: [items]
      properties:
        items: { type: array, maxItems: 50000, items: { $ref: "#/components/schemas/SitemapEntry" } }
    RedirectTargetResponse:
      type: object
      required: [target_path, status_code]
      properties:
        target_path: { type: string, pattern: "^/(works|performers|studios)/" }
        status_code: { const: 301 }
    WorkSearchResponse:
      type: object
      required: [query, items, next_cursor]
      properties:
        query:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/HomeItem"
        next_cursor: { type: [string, "null"], maxLength: 256 }
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    ItemListResponse:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/HomeItem" }
        next_cursor: { type: [string, "null"], maxLength: 256 }
...